:truck: [Utilities] Rename numbers.ts to number.ts
