+++ pkg/utilities/index.ts
@@ -1,4 +1,4 @@
 export {arrayOfLength, arrayShallowEquals} from './array';
 export {getErrorMessage} from './error';
-export {assertNumber, clamp, progressPercentage} from './numbers';
+export {assertNumber, clamp, progressPercentage} from './number';
 export {msToSec, secToMs} from './time';
