+++ pkg/utilities/number.ts
